various changes towards docker-compose/docker going
* create a webgui user to run webgui as.  probably not actually
needed but I was thrashing around trying to figure out why things
were barfing but I'll take it.  make docker-compose.yaml
su webgui -c plack.
* move docker-compose.yaml up.  maybe later I'll try to move
all of the build stuff to some subdir but I'm using it a lot
right now.
* add etc last in the docker recipe so config is the easiest
thing to change.
* use 'expose' instead of 'ports' in docker-compose for the
database because it only needs to connect to the other
container.
* add a webgui config file, etc/webgui.conf, and set a
DBI::mysql DSN that actually connects to the database,
which is db:3306.  hey, this mini-network docker-compose
stuff is pretty nifty!  some kind of DSN or something is
built in, so if a container is named db, I can refer to
it with the 'db' hostname from others, so that's what I
did.  also config has the correct password for the db.
* move various things to the share directory.  for install
files, share makes more sense than docs.  given enough install
files, eventually might move those to their own dir.
create.sql was very much an old-timey install file and it was
in share so following that lead as long as there aren't too
many.
* one stray doc comment.  I like doc comments.  the more I
have to unravel how parts fit together, the more doc comments
I write.