ongoing wG8 bootstrap inside docker work
needs more testing but initially and superficially, this might be
basically functional.

still needs nginx and logging configured, but upgrades run, and it
waits for them.  wgd db is supposed to fire up the database and
that's not working yet but if I can get it to go, it'll save having
to put the db username/dbname/password in the image twice.

added missing needed modules including modules needed to load
the initial upgrades needed to bootstrap 8.

wgd reset --uploads part of build process.

wgd reset --upgrade is run every time the wG container is started.
shouldn't hurt any thing.

wgd container now waits for the database to come online.
merely being able to connect to the db isn't good enough...
reading from the webguiVersion is a better indication that the
db is actually loaded though we could probably look for whichever
table loads absolutely last to really make sure that things are
probably loaded.

fix the uploadsPath.

all of these little things should be their own commit but I'm
thrashing around on getting things running.