entrypoint handling arbitrary wG conf files
That now uses `wgd db --print` to get database connection info from the config
file.

It now also only attempts upgrades if the code version is out of sync with the
database version, which should only happen when they pull a new version of wG
off of dockerhub or when a brand new 8 is first started.  I believe that gives
people control over upgrades.

It also loops over each config file found using `wgd for-each`.  If users
create a volume for `etc/` and use a different config file or multiple ones,
then everything will still work, and it'll upgrade each database that needs it.