use code from response->json({ code => }) as the HTTP status code
idiomatic shortcut to specifying an HTTP response code at the
same time as the JSON response body.
trying to keep JSON error responses to one line, with detecting
the AJAX request and returning the JSON error crammed in there.

related to  User list #45 and Auth module #41 and this whole
new-admin effort.