WebGUI::Fork fixes
child processes need to clone the connection and
the AutoInactiveDestroy connect param needs to be set if we
are going to fork or else the child process will disconnect
the parent's dbh when it exits.
WebGUI::Fork::init() now takes session but it's not clear that anything was ever calling init before.
it's a static method but WebGUI::Fork->init doesn't appear anywhere.
fixes to make t/Admin.t mostly work but $dbh is getting disconnected before WebGUI::Test goes
to cleanup and it's not clear why.