remove WebGUI::Asset::Wobject::WeatherData
leaving the wobject there but taking it out of the config.
broken assets prevent the admin from turning on (currently, with
Doug's admin).
there's no useful error to the user and a semi-helpful one to the
log.
the Weather::Com::Finder dep has a dep of its own that doesn't
build and that takes out the wobject.